Old/requirements.txt
discord.py==1.5.1
dnspython==2.0.0
pymongo==3.11.2
python-dotenv==0.15.0
requests-futures==1.0.0